+++ src/Fractions/Fraction.cs
@@ -23,7 +23,30 @@ public readonly partial struct Fraction :
 
 {
 #pragma warning disable IDE1006
-    private static readonly BigInteger TEN = new(10);
+    internal static readonly BigInteger TEN = new(10);
+
+    internal static readonly BigInteger[] PowersOfTen = [
+        BigInteger.One, // 10^0 = 1
+        new(10), // 10^1 = 10
+        new(100), // 10^2 = 100
+        new(1000), // 10^3 = 1,000
+        new(10000), // 10^4 = 10,000
+        new(100000), // 10^5 = 100,000
+        new(1000000), // 10^6 = 1,000,000
+        new(10000000), // 10^7 = 10,000,000
+        new(100000000), // 10^8 = 100,000,000
+        new(1000000000), // 10^9 = 1,000,000,000
+        new(10000000000), // 10^10 = 10,000,000,000
+        new(100000000000), // 10^11 = 100,000,000,000
+        new(1000000000000), // 10^12 = 1,000,000,000,000
+        new(10000000000000), // 10^13 = 10,000,000,000,000
+        new(100000000000000), // 10^14 = 100,000,000,000,000
+        new(1000000000000000), // 10^15 = 1,000,000,000,000,000
+        new(10000000000000000), // 10^16 = 10,000,000,000,000,000
+        new(100000000000000000), // 10^17 = 100,000,000,000,000,000
+        new(1000000000000000000) // 10^18 = 1,000,000,000,000,000,000
+    ];
+
 #pragma warning restore IDE1006
 
     private readonly BigInteger? _denominator;